port to `provider-interop`
In contrast with the current provider-concept, the `TwigServiceProvider` is now a proprietary Simplex provider, which uses the (`provider-interop`-adapted) Simplex API.

In the particular case of Pimple/Simplex, there is no provider-concept native to those containers, so the provider-implementation here is just a class, it doesn't implement a particular interface - many other container-implementations (such as PHP-DI, Unbox, etc.) include a proprietary provider-concept, and the "Simplex-provider" implemented here is proprietary to Simplex as well, due to the type-hinted `bootstrap()` method and the use of the Simplex API.

Bottom line, the provider implementation itself is proprietary - the Simplex container implements the `ServiceProviderInterface`, which would enable other containers to import entries from the bootstrapped Simplex `Container`.

To contrast with the original service-provider proposal, interoperability is now provided by the container-implementation, not by the provider, which remains proprietary and can freely leverage all the features of the DI container selected by the module vendor.

To have a functioning Twig environment, there are a few hard dependencies (template and cache paths) which were previously hidden - these are now visible and required by the constructor, making it necessary for the developer to provide the required information up front; we don't bootstrap and then patch the bootstrapping later, instead we make the required decisions up front.

Also note that all the methods are now non-`static`, such that the provider can be extended, and internal methods are `protected`, such that internal methods don't leave a public API footprint.